feat(npm): drop embedded managers
Renovate will no longer embed npm, yarn and pnpm as dependencies. Therefore, if you are not running via the Docker full image, then you need to ensure that any of these managers are installed prior to running Renovate. This is inline with other managers such as Gradle or Poetry that require installation alongside Renovate too.

Closes #4208

BREAKING CHANGE: npm, yarn and pnpm are no longer embedded with the installation